+++ Data_formatting_scripts/make_urma_datasets.sh
@@ -53,15 +53,15 @@ for YYYYMMDD in *; do
     cd ${YYYYMMDD}
     for file in *; do
         if [[ ! "${file}" == *"idx" ]]; then #make sure we avoid the grib2 index files, if they exist
+            # Extract just the tXXz part of file name, assuming it looks like [something].tXXz.[something].grib2
+            tXXz=${file#*.}
+            tXXz=${tXXz%.*}
+            tXXz=${tXXz%.*}
+            
             for IDX in "${!VAR_LIST[@]}"; do
-                # Extract just the tXXz part of file name, assuming it looks like [something].tXXz.[something].grib2
-                tXXz=${file#*.}
-                tXXz=${tXXz%.*}
-                tXXz=${tXXz%.*}
-
                 target_filename="urma_${VAR_LIST[$IDX]}_${YYYYMMDD}_${tXXz}_regridded.grib2"
 
-                if [ ! -f "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/${target_filename}" ]; then #file DNE in target directory, safe to proceed
+                if [[ ! -e "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/${target_filename}" ]]; then #file DNE in target directory, safe to proceed
                     wgrib2 ${file} -match "${VAR_SELECTION_LIST[$IDX]}" -grib "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/TEMP.grib2" #subset the appropriate variable into a temporary holding file
                     wgrib2 "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/TEMP.grib2" -set_radius 1:6370000 -grib_out "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/${target_filename}"
                     target_filesize=$(stat -c '%s' "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/${target_filename}")
@@ -72,12 +72,13 @@ for YYYYMMDD in *; do
                         wgrib2 "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/TEMP.grib2" -set_radius 1:6370000 -grib_out "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/${target_filename}"
                         target_filesize=$(stat -c '%s' "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/${target_filename}")
                     done #end filesize check and while loop
+                    echo "${target_filename} DONE"
 
                 else #if file exists, make sure it's not tiny, which happens often if process is interrupted mid-run
                     target_filesize=$(stat -c '%s' "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/${target_filename}")
-                    echo "${target_filename} | filesize = ${target_filesize} | REMAKING!!"
                     while (( target_filesize < 10000 )); do 
-                        if [ -f "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/TEMP.grib2" ]; then
+                        echo "${target_filename} | filesize = ${target_filesize} | REMAKING!!"
+                        if [[ -e "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/TEMP.grib2" ]]; then
                             rm "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/TEMP.grib2"
                         fi
                         rm "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/${target_filename}"
@@ -86,7 +87,11 @@ for YYYYMMDD in *; do
                         target_filesize=$(stat -c '%s' "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/${target_filename}")
                     done #end filesize check and while loop
                 fi #end target_filename check
-                rm "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/TEMP.grib2"
+
+                if [[ -e "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/TEMP.grib2" ]]; then
+                    rm "${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}/TEMP.grib2"
+                fi
+                
             done #end "for IDX in "${!VAR_LIST[@]}" "
         fi #end 'if [ ( ! "${file}" == *"idx" ) && ( "${file}" == *"wrfnatf01"* ) ]'
     done #end 'for file in *'
@@ -96,5 +101,25 @@ done #end 'for YYYYMMDD in *'
 ########################################################################
 ### Concatenate into training and testing datasets 
 
-### TO DO once all the data has been generated
-## No need for date checks here, just put all 2021-23 data into training and 2024 into testing
+for IDX in "${!VAR_LIST[@]}"; do
+    cd ${URMA_REGRIDDED_DATA_DIR}/${VAR_LIST[$IDX]}
+
+    echo "Making URMA training dataset for ${VAR_LIST[$IDX]} (2021/22/23)..."
+    output_filename=train_urma_alltimes_CONUS_${VAR_LIST[$IDX]}.grib2
+    if [[ ! -e ${URMA_TRAIN_TEST_DIR}/${output_filename} ]]; then
+        cat *_2021*.grib2 *_2022*.grib2 *_2023*.grib2 > ${URMA_TRAIN_TEST_DIR}/${output_filename}
+        echo "URMA training dataset for ${VAR_LIST[$IDX]} is done"
+    else
+        echo "!! URMA training dataset for ${VAR_LIST[$IDX]} already exists at ${URMA_TRAIN_TEST_DIR}/${output_filename}"
+    fi
+
+    echo "Making URMA testing dataset for ${VAR_LIST[$IDX]} (2024)..."
+    output_filename=test_urma_alltimes_CONUS_${VAR_LIST[$IDX]}.grib2
+    if [[ ! -e ${URMA_TRAIN_TEST_DIR}/${output_filename} ]]; then
+        cat *_2024*.grib2 > ${URMA_TRAIN_TEST_DIR}/${output_filename}
+        echo "URMA testing dataset for ${VAR_LIST[$IDX]} is done"
+    else
+        echo "!! URMA testing dataset for ${VAR_LIST[$IDX]} already exists at ${URMA_TRAIN_TEST_DIR}/${output_filename}"
+    fi
+    
+done #end loop over VAR_LIST
